Split BoundaryValuesFactory into compilation unit
